+++ futurehome/src/index.ts
@@ -7,7 +7,7 @@ import { haUpdateState, haUpdateStateValueReport } from './ha/update_state';
 import { VinculumPd7Device } from './fimp/vinculum_pd7_device';
 import { haUpdateAvailability } from './ha/update_availability';
 import { delay } from './utils';
-import { exposeSmarthubTools, handleInclusionStatusReport } from './ha/admin';
+import { exposeSmarthubTools, handleExclusionReport, handleExclusionStatusReport, handleInclusionStatusReport } from './ha/admin';
 import { pollVinculum } from './fimp/vinculum';
 
 (async () => {
@@ -146,8 +146,11 @@ import { pollVinculum } from './fimp/vinculum';
         const result = haPublishDevice({
           hubId,
           demoMode,
+          hubIp,
           vinculumDeviceData,
           deviceInclusionReport,
+          thingsplexUsername,
+          thingsplexPassword,
         });
         await delay(50);
 
@@ -253,6 +256,16 @@ import { pollVinculum } from './fimp/vinculum';
 
         case 'evt.thing.inclusion_status_report': {
           handleInclusionStatusReport(hubId, msg);
+          break;
+        }
+        
+        case 'evt.thing.exclusion_status_report': {
+          handleExclusionStatusReport(hubId, msg);
+          break;
+        }
+        
+        case 'evt.thing.exclusion_report': {
+          handleExclusionReport();
           break;
         }