+++ futurehome/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added logo.
 - Added support for pairing new devices.
+- Added support for unpairing devices.
 
 ## 0.1.7 (26.07.2025)